Remove old Protocols and Codecs
