--- client/hotrod-client/src/main/java/org/infinispan/client/hotrod/impl/protocol/Codec25.java
@@ -1,20 +1,0 @@
-package org.infinispan.client.hotrod.impl.protocol;
-
-import io.netty.buffer.ByteBuf;
-
-/**
- * @author gustavonalle
- * @since 8.2
- */
-public class Codec25 extends Codec24 {
-
-   @Override
-   public HeaderParams writeHeader(ByteBuf buf, HeaderParams params) {
-      return writeHeader(buf, params, HotRodConstants.VERSION_25);
-   }
-
-   @Override
-   public short readMeta(ByteBuf buf) {
-      return buf.readUnsignedByte();
-   }
-}
