--- client/hotrod-client/src/test/java/org/infinispan/client/hotrod/query/RemoteQueryDslConditionsOldClient.java
@@ -1,14 +1,0 @@
-package org.infinispan.client.hotrod.query;
-
-import org.infinispan.client.hotrod.ProtocolVersion;
-import org.testng.annotations.Test;
-
-
-@Test(groups = "functional", testName = "client.hotrod.query.RemoteQueryDslConditionsOldClient")
-public class RemoteQueryDslConditionsOldClient extends RemoteQueryDslConditionsTest {
-
-   @Override
-   protected ProtocolVersion getProtocolVersion() {
-      return ProtocolVersion.PROTOCOL_VERSION_27;
-   }
-}
