--- integrationtests/jboss-marshalling-it/src/test/java/org/infinispan/it/marshalling/jboss/JBMARRemoteQueryOldClientDslConditionsTest.java
@@ -1,13 +1,0 @@
-package org.infinispan.it.marshalling.jboss;
-
-import org.infinispan.client.hotrod.ProtocolVersion;
-import org.testng.annotations.Test;
-
-@Test(groups = "functional", testName = "client.hotrod.query.JBMARRemoteQueryOldClientDslConditionsTest")
-public class JBMARRemoteQueryOldClientDslConditionsTest extends JBMARRemoteQueryDslConditionsTest {
-
-   @Override
-   protected ProtocolVersion getProtocolVersion() {
-      return ProtocolVersion.PROTOCOL_VERSION_27;
-   }
-}
