+++ spring/spring-boot-3/remote/src/test/resources/custom-test-hotrod-client.properties
@@ -7,7 +7,7 @@ infinispan.client.hotrod.socket_timeout=3000
 infinispan.client.hotrod.connect_timeout=5000
 infinispan.client.hotrod.max_retries=42
 infinispan.client.hotrod.batch_size=90
-infinispan.client.hotrod.protocol_version=2.8
+infinispan.client.hotrod.protocol_version=4.0
 
 # connection pool
 infinispan.client.hotrod.connection_pool.max_active=91
