+++ spring/spring-boot-3/remote/src/test/resources/test-application.properties
@@ -7,7 +7,7 @@ infinispan.remote.socket-timeout=500
 infinispan.remote.connect-timeout=200
 infinispan.remote.max-retries=30
 infinispan.remote.batch-size=91
-infinispan.remote.protocol-version=2.4
+infinispan.remote.protocol-version=3.1
 
 # pool
 infinispan.remote.connection-pool-max-active=90
